feat(api): integrate quiz and flashcard services with HTTP endpoints
- Update dependency injection to wire QuizService and FlashcardService
- Refactor quiz router: cleaner session and question endpoints with Graph RAG
- Update flashcards router for study-progress integration
- Add Pydantic schemas for quiz and flashcard API contracts
- Enforce client-side response redaction: never expose sensitive fields
- Maintain API backward compatibility with existing endpoints

+++ mindforge/api/schemas.py
@@ -175,7 +175,7 @@ class AnswerEvaluationResponse(BaseModel):
     """Evaluation result returned to client — reference_answer NOT included."""
 
     question_id: str
-    score: float
+    score: int
     feedback: str
     is_correct: bool
